bugs arrumados
