+++ pokedex-labenu/src/pages/home/style.js
@@ -1,19 +1,7 @@
 import styled from "styled-components";
-import { AppBar, Toolbar } from "@material-ui/core";
+
 // import { theme } from "../../themes/index";
 
-export const HeaderAppBar = styled(AppBar)`
-  background-color: ${(props) => props.theme.main};
-`;
-export const HeaderToolbar = styled(Toolbar)`
-  display: flex;
-  justify-content: space-between;
-  gap: 16px;
-  flex-wrap: wrap;
-`;
-export const HeaderLogo = styled.img`
-  width: 3vw;
-`;
 export const CardsContainer = styled.div`
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
